refactor(http)!: only expose ratelimit info on response error
The other variants were only useful prior to the API V8 error changes.
