+++ http/src/lib.rs
@@ -128,7 +128,6 @@
     clippy::unnecessary_wraps
 )]
 
-pub mod api_error;
 pub mod client;
 pub mod error;
 pub mod request;
